Generatable interface has been rewritten to produce larger file.

+++ src/com/d9nich/generator/SequentGenerator.java
@@ -1,6 +1,9 @@
 package com.d9nich.generator;
 
-import java.io.*;
+import java.io.BufferedOutputStream;
+import java.io.DataOutputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
 
 public class SequentGenerator implements Generatable {
 
@@ -9,7 +12,7 @@ public class SequentGenerator implements Generatable {
      * @param dataSize how many integers we should write to file
      */
     @Override
-    public void generate(String fileName, int dataSize) {
+    public void generate(String fileName, long dataSize) {
         int number = Integer.MIN_VALUE;
         try (DataOutputStream output = new DataOutputStream(new BufferedOutputStream(new FileOutputStream(fileName)))) {
             for (int i = 0; i < dataSize; i++)
