+++ src/com/d9nich/generator/SmallGenerator.java
@@ -12,7 +12,7 @@ public class SmallGenerator implements Generatable {
      * @param dataSize not used only to support interface
      */
     @Override
-    public void generate(String fileName, int dataSize) {
+    public void generate(String fileName, long dataSize) {
         int[] numbers = {2, 3, 4, 0, 5, 6, 7, 9, 8, 1};
         try (DataOutputStream output = new DataOutputStream(new BufferedOutputStream(new FileOutputStream(fileName)))) {
             for (int number : numbers)
